+++ src/DeployTest/opencl_utils.h
@@ -27,7 +27,7 @@ extern std::vector<device> devices;
 
 const uint8_t Init();
 const std::string DeviceTypeString(const cl_device_type type);
-const uint8_t GetRecommendedDevices(const uint8_t count, std::vector<device> &devices);
+const uint8_t GetRecommendedDevices(const uint16_t count, std::vector<device> &devices);
 const void PrintInfo();
 }
